passing hf manager tests
