+++ tests/trainers/conftest.py
@@ -152,3 +152,7 @@ class MockTrainer(BaseTrainer):
 
     def evaluate(self) -> TestResult:
         return TestResult(loss=0.42, metrics={"metric_1": 42})
+
+    @property
+    def model(self) -> str:
+        return "mock model"
